Refactor TwoFactorAuthenticatorsController#create
